RaitingManagment/requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.5
flask-mongoengine==1.0.0
Flask-MySQLdb==0.2.0
Flask-WTF==1.2.2
google-auth==2.36.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mongoengine==0.29.1
mysqlclient==2.2.6
pika==1.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
